pyproject.toml
[project]
name = "02-real-estate-ai-vscode"
version = "0.1.0"
description = ""
authors = [
    {name = "Takumi Yamamoto",email = "yama.moomin08@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "fastapi[all] (>=0.116.0,<0.117.0)",
    "openai (>=1.93.3,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"